i2c: add possibility to create slave without init
Upstream i2c.c has one method for creating an i2c slave device instance,
the i2c_create_slave function. This method creates the qdev instance and
calls init for it so there is no room for setting any device specific
properties that should be filled in between device creation and init. This
patch adds another method, i2c_create_slave_noinit, which does the same as
the existing method but leaves the returned device uninitialized. Caller is
responsible for initializing it.

Signed-off-by: Juha Riihimäki <juha.riihimaki@nokia.com>