linux-user: Dummy implementation of prlimit64 to avoid warnings
Special-case prlimit64 to return ENOSYS without printing the default
"unhandled syscall" warning; this is a new syscall so anything trying
to use it will have a fallback case, and the warning can result in
a lot of noise (for instance, when running apt under qemu).

This will be replaced with a proper implementation later.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>