add generic qdev callback for monitor "change" command
When monitor command "change" is executed for a device which is not
named "vnc" nor a block-device, the main system bus is searched for
the specified device and if found, a generic qdev callback function
will be called to let the device react to the command in a specific
way.

TODO: the use of this is in lis302dl in hw/nseries.c, added
in the "add n900" patch. Consider splitting it out as controversial?

Signed-off-by: Juha Riihimäki <juha.riihimaki@nokia.com>