hw/pxa2xx.c: Fix handling of pxa2xx_i2c variable offset within region
The pxa2xx I2C controller can be at an arbitrary offset within its
region (this is used because one of the controllers starts at offset
0x1600 into an 0x10000 sized region). The previous implementation of this
included an adjustment which worked around the fact that memory region
read/write functions were passed an offset from the start of a page
rather than from the start of the region. Since commit 5312bd8b3
offsets are now from the start of the region and so we were applying
an incorrect adjustment, resulting in warnings like
"pxa2xx_i2c_read: Bad register 0xffffff90".

Retain the offset handling but remove the adjustment to the page
boundary.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>