arm: kvm: Fix register sync and some GDB support
The register copies between the kernel and QEMU did not properly
consider QEMUs mode of putting all current registers in a given array
regardless of the mode and at the same time having banked registers for
the same registers. With this commit we should be sure to be in sync
always.

Also introduces a little bit of GDB support. QEMU does indeed translate
virtual addresses to physical by reading guest memory directly in order
to access guest memory for debugging, so guest LPAE support with GDB
would rely on having the LPAE page table format supported in QEMU.

Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>