hw/omap_tap.c: Return correct ID register values for OMAP3
Add/correct the ID register values returned for OMAP3.
In particular, the OMAP3 PRODUCTION_ID register was not
modelled correctly: it is a 128 bit register spanning
four words 0x208..0x214. (There appears to be a typo in
the OMAP35xx TRM about this; however the OMAP36xx TRM
is clearer and experimentation shows that the 35xx behaves the
same way.)

Signed-off-by:  Peter Maydell <peter.maydell@linaro.org>