hw/omap_uart.c: Refactor register access mode tests
The OMAP UART provides different views of the registers depending on the
access mode specified by the value in the LCR register. Refactor this
to use an enumeration indicating the current mode rather than doing
checks of the cache LCR value everywhere. Similarly, pull out into a
function the check of whether registers 0x18/0x1c are in TCR/TLR mode.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>